Fix rocmlir after rebase

+++ mlir/test/Conversion/TosaToRock/tosa-to-rock-attention.mlir
@@ -37,12 +37,12 @@ func.func @self_attention_with_bias_only(%arg0: tensor<1x384x64xf32>, %arg1: ten
   %0 = "tosa.transpose"(%arg1, %cst) : (tensor<1x384x64xf32>, tensor<3xi64>) -> tensor<1x64x384xf32>
   %1 = "tosa.matmul"(%arg0, %0) : (tensor<1x384x64xf32>, tensor<1x64x384xf32>) -> tensor<1x384x384xf32>
   %2 = "tosa.add"(%1, %arg3) : (tensor<1x384x384xf32>, tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
-  %3 = "tosa.reduce_max"(%2) {axis = 1 : i64} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
+  %3 = "tosa.reduce_max"(%2) {axis = 1 : i32} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
   %4 = "tosa.sub"(%2, %3) : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
   %5 = "tosa.exp"(%4) : (tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
-  %6 = "tosa.reduce_sum"(%5) {axis = 1 : i64} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
+  %6 = "tosa.reduce_sum"(%5) {axis = 1 : i32} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
   %7 = "tosa.reciprocal"(%6) : (tensor<1x1x384xf32>) -> tensor<1x1x384xf32>
-  %8 = "tosa.mul"(%5, %7) {shift = 0 : i32} : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
+  %8 = "tosa.mul"(%5, %7) {shift = 0 : i8} : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
   %9 = "tosa.matmul"(%8, %arg2) : (tensor<1x384x384xf32>, tensor<1x384x64xf32>) -> tensor<1x384x64xf32>
   return %9 : tensor<1x384x64xf32>
 }
@@ -52,14 +52,14 @@ func.func @self_attention_with_scale_and_bias(%arg0: tensor<1x384x64xf32>, %arg1
   %cst = arith.constant dense<[0, 2, 1]> : tensor<3xi64>
   %0 = "tosa.transpose"(%arg1, %cst) : (tensor<1x384x64xf32>, tensor<3xi64>) -> tensor<1x64x384xf32>
   %1 = "tosa.matmul"(%arg0, %0) : (tensor<1x384x64xf32>, tensor<1x64x384xf32>) -> tensor<1x384x384xf32>
-  %2 = "tosa.mul"(%1, %arg3) {shift = 0 : i32} : (tensor<1x384x384xf32>, tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
+  %2 = "tosa.mul"(%1, %arg3) {shift = 0 : i8} : (tensor<1x384x384xf32>, tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
   %3 = "tosa.add"(%2, %arg4) : (tensor<1x384x384xf32>, tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
-  %4 = "tosa.reduce_max"(%3) {axis = 1 : i64} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
+  %4 = "tosa.reduce_max"(%3) {axis = 1 : i32} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
   %5 = "tosa.sub"(%3, %4) : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
   %6 = "tosa.exp"(%5) : (tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
-  %7 = "tosa.reduce_sum"(%6) {axis = 1 : i64} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
+  %7 = "tosa.reduce_sum"(%6) {axis = 1 : i32} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
   %8 = "tosa.reciprocal"(%7) : (tensor<1x1x384xf32>) -> tensor<1x1x384xf32>
-  %9 = "tosa.mul"(%6, %8) {shift = 0 : i32} : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
+  %9 = "tosa.mul"(%6, %8) {shift = 0 : i8} : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
   %10 = "tosa.matmul"(%9, %arg2) : (tensor<1x384x384xf32>, tensor<1x384x64xf32>) -> tensor<1x384x64xf32>
   return %10 : tensor<1x384x64xf32>
 }
@@ -69,15 +69,15 @@ func.func @self_attention_with_scale_bias_exp(%arg0: tensor<1x384x64xf32>, %arg1
   %cst = arith.constant dense<[0, 2, 1]> : tensor<3xi64>
   %0 = "tosa.transpose"(%arg1, %cst) : (tensor<1x384x64xf32>, tensor<3xi64>) -> tensor<1x64x384xf32>
   %1 = "tosa.matmul"(%arg0, %0) : (tensor<1x384x64xf32>, tensor<1x64x384xf32>) -> tensor<1x384x384xf32>
-  %2 = "tosa.mul"(%1, %arg3) {shift = 0 : i32} : (tensor<1x384x384xf32>, tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
+  %2 = "tosa.mul"(%1, %arg3) {shift = 0 : i8} : (tensor<1x384x384xf32>, tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
   %3 = "tosa.add"(%2, %arg4) : (tensor<1x384x384xf32>, tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
   %exp = "tosa.exp"(%3) : (tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
-  %4 = "tosa.reduce_max"(%exp) {axis = 1 : i64} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
+  %4 = "tosa.reduce_max"(%exp) {axis = 1 : i32} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
   %5 = "tosa.sub"(%exp, %4) : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
   %6 = "tosa.exp"(%5) : (tensor<1x384x384xf32>) -> tensor<1x384x384xf32>
-  %7 = "tosa.reduce_sum"(%6) {axis = 1 : i64} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
+  %7 = "tosa.reduce_sum"(%6) {axis = 1 : i32} : (tensor<1x384x384xf32>) -> tensor<1x1x384xf32>
   %8 = "tosa.reciprocal"(%7) : (tensor<1x1x384xf32>) -> tensor<1x1x384xf32>
-  %9 = "tosa.mul"(%6, %8) {shift = 0 : i32} : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
+  %9 = "tosa.mul"(%6, %8) {shift = 0 : i8} : (tensor<1x384x384xf32>, tensor<1x1x384xf32>) -> tensor<1x384x384xf32>
   %10 = "tosa.matmul"(%9, %arg2) : (tensor<1x384x384xf32>, tensor<1x384x64xf32>) -> tensor<1x384x64xf32>
   return %10 : tensor<1x384x64xf32>
 }
